+++ dev/src/crossroads.js
@@ -4,24 +4,28 @@
     
     crossroads = (function(){
         
-        var _crossroads = {
+        function Crossroads(){
+            this._routes = [];
+            this.bypassed = new signals.Signal();
+            this.routed = new signals.Signal();
+        }
+
+        Crossroads.prototype = {
             
+            create : function(){
+                return new Crossroads();
+            },
+
             shouldTypecast : true,
 
-            _routes : [],
-            
-            bypassed : new signals.Signal(),
-            
-            routed : new signals.Signal(),
-
             addRoute : function(pattern, callback, priority){
-                var route = new Route(pattern, callback, priority);
-                sortedInsert(route);
+                var route = new Route(pattern, callback, priority, this);
+                this._sortedInsert(route);
                 return route;
             },
             
             removeRoute : function(route){
-                var i = getRouteIndex(route);
+                var i = arrayIndexOf(this._routes, route);
                 if(i >= 0) this._routes.splice(i, 1);
                 route._destroy();
             },
@@ -36,8 +40,8 @@
             
             parse : function(request){
                 request = request || '';
-                var route = getMatchedRoute(request),
-                    params = route? getParamValues(request, route) : null;
+                var route = this._getMatchedRoute(request),
+                    params = route? patternLexer.getParamValues(request, route._matchRegexp, this.shouldTypecast) : null;
                 if(route){
                     params? route.matched.dispatch.apply(route.matched, params) : route.matched.dispatch();
                     this.routed.dispatch(request, route, params);
@@ -50,38 +54,30 @@
                 return this._routes.length;
             },
 
+            _sortedInsert : function (route){
+                //simplified insertion sort
+                var routes = this._routes,
+                    n = routes.length;
+                do { --n; } while (routes[n] && route._priority <= routes[n]._priority);
+                routes.splice(n+1, 0, route);
+            },
+            
+            _getMatchedRoute : function (request){
+                var routes = this._routes,
+                    n = routes.length,
+                    route;
+                while(route = routes[--n]){ //should be decrement loop since higher priorities are added at the end of array  
+                    if(route.match(request)) return route;
+                }
+                return null;
+            },
+
             toString : function(){
                 return '[crossroads numRoutes:'+ this.getNumRoutes() +']';
             }
         };
         
-        function sortedInsert(route){
-            //simplified insertion sort
-            var routes = crossroads._routes,
-                n = routes.length;
-            do { --n; } while (routes[n] && route._priority <= routes[n]._priority);
-            routes.splice(n+1, 0, route);
-        }
-        
-        function getRouteIndex(route){
-            return arrayIndexOf(crossroads._routes, route);
-        }
-        
-        function getMatchedRoute(request){
-            var routes = crossroads._routes,
-                n = routes.length,
-                route;
-            while(route = routes[--n]){ //should be decrement loop since higher priorities are added at the end of array  
-                if(route.match(request)) return route;
-            }
-            return null;
-        }
-        
-        function getParamValues(request, route){
-            return patternLexer.getParamValues(request, route._matchRegexp);
-        }
-        
-        //API
-        return _crossroads;
+        //return "static" instance
+        return new Crossroads();
         
     }());