+++ CHANGELOG.md
@@ -8,6 +8,7 @@
  - added property `crossroads.shouldTypecast` to enable/disable typecasting
    segments values. (issue #18)
  - added support to optional segments (issue #17)
+ - added support to multiple instances (issue #19)
 
 ### Other ###